updated: moved the browser/index.js into riot.js in favore of a simpler export strategy
