+++ lib/riot.js
@@ -1,21 +1,216 @@
-import riot from './browser/index'
-export default riot
+import observable from 'riot-observable'
+import { tmpl, brackets } from 'riot-tmpl'
 
-/*
-import 'browser/wrap/prefix'
-import 'browser/global-variables'
-import '../node_modules/riot-observable/dist/riot.observable.js'
-import '../node_modules/riot-tmpl/dist/riot.tmpl.js'
-import 'browser/tag/mkdom'
-import 'browser/tag/each'
-import 'browser/tag/if'
-import 'browser/tag/named'
-import 'browser/tag/styleManager'
-import 'browser/tag/parse'
-import 'browser/tag/tag'
-import 'browser/tag/update'
-import 'browser/tag/util'
-import 'browser/index'
-import 'browser/wrap/suffix'
-*/
+import { isFunction, setAttr, $$, each } from './browser/util'
+import { __tagImpl, __virtualDom, T_STRING } from './browser/global-variables'
+import styleManager from './browser/tag/styleManager'
+import Tag from './browser/tag/tag'
 
+/**
+ * Riot public api
+ */
+
+export var observable = observable
+
+// share methods for other riot parts, e.g. compiler
+export var util = { brackets: brackets, tmpl: tmpl }
+
+/**
+ * Create a mixin that could be globally shared across all the tags
+ */
+export var mixin = (function() {
+  var mixins = {},
+    globals = mixins[GLOBAL_MIXIN] = {},
+    _id = 0
+
+  /**
+   * Create/Return a mixin by its name
+   * @param   { String }  name - mixin name (global mixin if object)
+   * @param   { Object }  mixin - mixin logic
+   * @param   { Boolean } g - is global?
+   * @returns { Object }  the mixin logic
+   */
+  return function(name, mixin, g) {
+    // Unnamed global
+    if (isObject(name)) {
+      riot.mixin('__unnamed_'+_id++, name, true)
+      return
+    }
+
+    var store = g ? globals : mixins
+
+    // Getter
+    if (!mixin) return store[name]
+    // Setter
+    store[name] = extend(store[name] || {}, mixin)
+  }
+
+})()
+
+/**
+ * Create a new riot tag implementation
+ * @param   { String }   name - name/id of the new riot tag
+ * @param   { String }   html - tag template
+ * @param   { String }   css - custom tag css
+ * @param   { String }   attrs - root tag attributes
+ * @param   { Function } fn - user function
+ * @returns { String } name/id of the tag just created
+ */
+export function tag(name, html, css, attrs, fn) {
+  if (isFunction(attrs)) {
+    fn = attrs
+    if (/^[\w\-]+\s?=/.test(css)) {
+      attrs = css
+      css = ''
+    } else attrs = ''
+  }
+  if (css) {
+    if (isFunction(css)) fn = css
+    else styleManager.add(css)
+  }
+  name = name.toLowerCase()
+  __tagImpl[name] = { name: name, tmpl: html, attrs: attrs, fn: fn }
+  return name
+}
+
+/**
+ * Create a new riot tag implementation (for use by the compiler)
+ * @param   { String }   name - name/id of the new riot tag
+ * @param   { String }   html - tag template
+ * @param   { String }   css - custom tag css
+ * @param   { String }   attrs - root tag attributes
+ * @param   { Function } fn - user function
+ * @returns { String } name/id of the tag just created
+ */
+export function tag(name, html, css, attrs, fn) {
+  if (css) styleManager.add(css)
+  //if (bpair) riot.settings.brackets = bpair
+  __tagImpl[name] = { name: name, tmpl: html, attrs: attrs, fn: fn }
+  return name
+}
+
+/**
+ * Mount a tag using a specific tag implementation
+ * @param   { String } selector - tag DOM selector
+ * @param   { String } tagName - tag implementation name
+ * @param   { Object } opts - tag logic
+ * @returns { Array } new tags instances
+ */
+export function mount(selector, tagName, opts) {
+
+  var els,
+    allTags,
+    tags = []
+
+  // helper functions
+
+  function addRiotTags(arr) {
+    var list = ''
+    each(arr, function (e) {
+      if (!/[^-\w]/.test(e)) {
+        e = e.trim().toLowerCase()
+        list += ',[' + RIOT_TAG_IS + '="' + e + '"],[' + RIOT_TAG + '="' + e + '"]'
+      }
+    })
+    return list
+  }
+
+  function selectAllTags() {
+    var keys = Object.keys(__tagImpl)
+    return keys + addRiotTags(keys)
+  }
+
+  function pushTags(root) {
+    if (root.tagName) {
+      var riotTag = getAttr(root, RIOT_TAG_IS) || getAttr(root, RIOT_TAG)
+
+      // have tagName? force riot-tag to be the same
+      if (tagName && riotTag !== tagName) {
+        riotTag = tagName
+        setAttr(root, RIOT_TAG_IS, tagName)
+        setAttr(root, RIOT_TAG, tagName) // this will be removed in riot 3.0.0
+      }
+      var tag = mountTo(root, riotTag || root.tagName.toLowerCase(), opts)
+
+      if (tag) tags.push(tag)
+    } else if (root.length) {
+      each(root, pushTags)   // assume nodeList
+    }
+  }
+
+  // ----- mount code -----
+
+  // inject styles into DOM
+  styleManager.inject()
+
+  if (isObject(tagName)) {
+    opts = tagName
+    tagName = 0
+  }
+
+  // crawl the DOM to find the tag
+  if (typeof selector === T_STRING) {
+    if (selector === '*')
+      // select all the tags registered
+      // and also the tags found with the riot-tag attribute set
+      selector = allTags = selectAllTags()
+    else
+      // or just the ones named like the selector
+      selector += addRiotTags(selector.split(/, */))
+
+    // make sure to pass always a selector
+    // to the querySelectorAll function
+    els = selector ? $$(selector) : []
+  }
+  else
+    // probably you have passed already a tag or a NodeList
+    els = selector
+
+  // select all the registered and mount them inside their root elements
+  if (tagName === '*') {
+    // get all custom tags
+    tagName = allTags || selectAllTags()
+    // if the root els it's just a single tag
+    if (els.tagName)
+      els = $$(tagName, els)
+    else {
+      // select all the children for all the different root elements
+      var nodeList = []
+      each(els, function (_el) {
+        nodeList.push($$(tagName, _el))
+      })
+      els = nodeList
+    }
+    // get rid of the tagName
+    tagName = 0
+  }
+
+  pushTags(els)
+
+  return tags
+}
+
+/**
+ * Update all the tags instances created
+ * @returns { Array } all the tags instances
+ */
+export function update() {
+  return each(__virtualDom, function(tag) {
+    tag.update()
+  })
+}
+
+export function unregister(name) {
+  delete __tagImpl[name]
+}
+
+/**
+ * Export the Virtual DOM
+ */
+export var vdom = __virtualDom
+
+/**
+ * Export the Tag constructor
+ * TODO: make a better tag constructor
+ */
+// export function Tag() {}
